feat: add geofence check for arbitrary geometry objects
It's now possible to check if a arbitrary geometry object lies inside
a polygon (until now this check was only possible for single points).
A `Geometry` interface is introduced with this commit: `Coordinate`,
`Line`, `Polyline`, and `Polygon` implement that interface.

The `Polygon` class has a new method `containsGeometry()` which checks
for every point in a geometry instance if it's inside the polygon. The
restrictions for the `contains()` method apply.